# Move perf test command to separate subcommand