--- a/.github/in-cluster-test-scripts/aks.sh
+++ b/.github/in-cluster-test-scripts/aks.sh
@@ -19,7 +19,7 @@ sleep 10s
 cilium connectivity test --debug --all-flows --collect-sysdump-on-failure --external-target bing.com
 
 # Run performance test
-cilium connectivity test --perf --perf-duration 1s
+cilium connectivity perf --duration 1s
 
 # Retrieve Cilium status
 cilium status