--- a/.github/in-cluster-test-scripts/eks.sh
+++ b/.github/in-cluster-test-scripts/eks.sh
@@ -31,4 +31,4 @@ sleep 10s
 cilium connectivity test --debug --all-flows --collect-sysdump-on-failure --external-target amazon.com
 
 # Run performance test
-cilium connectivity test --perf --perf-duration 1s
+cilium connectivity perf --duration 1s