I really tried to be mindful of splitting that into commits, but unfortunately, I ended up refactoring the whole thing :cry: 

Summary:
- Move `cilium connectivity test --perf` to separate subcommand `cilium connectivity perf`
- Allow to run host/pod tests together (before you could only run one of), while also allow to test host-to-pod, pod-to-host etc.
- Before NodeSelector was totally not working in perf test and we were forcing users to use one zone for nodes, now nodeSelector works and we just warn in case nodes are in different zones (so now we can test both in-zone and across-zones)
-  Added possibility to export results in json format that can be visualized with perfdash - this will be used for tracking datapath performance in CI
- Visualize operation rate next to latency in console output (as they were from the same netperf test)
- Always cleanup perf pods before performing tests - to make sure we respect nodeSelector
- Refactored netperf integration, it was full of hacks.

Help for perf subcommand:
```
cilium connectivity perf --help
Test network performance

Usage:
  cilium connectivity perf [flags]

Flags:
  -d, --debug                             Show debug messages
      --deployment-pod-annotations json   Add annotations to the connectivity pods, e.g. '{"client":{"foo":"bar"}}' (default {})
      --duration duration                 Duration for the Performance test to run (default 10s)
  -h, --help                              help for perf
      --host-net                          Test host network
      --node-selector stringToString      Restrict connectivity pods to nodes matching this label (default [])
      --performance-image string          Image path to use for performance (default "quay.io/cilium/network-perf:a816f935930cb2b40ba43230643da4d5751a5711@sha256:679d3a370c696f63884da4557a4466f3b5569b4719bb4f86e8aac02fbe390eea")
      --pod-net                           Test pod network (default true)
      --report-dir string                 Directory to save perf results in json format
      --samples int                       Number of Performance samples to capture (how many times to run each test) (default 1)
      --test-namespace string             Namespace to perform the connectivity in (default "cilium-test")

Global Flags:
      --context string     Kubernetes configuration context
  -n, --namespace string   Namespace Cilium is running in (default "kube-system")

```

Example output of full tests:
```
🔥 Network Performance Test Summary:
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
📋 Scenario        | Node       | Test            | Duration        | Min             | Mean            | Max             | P50             | P90             | P99             | Transaction rate OP/s
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
📋 pod-to-pod      | same-node  | TCP_RR          | 1s              | 16µs            | 36.6µs          | 6.983ms         | 22µs            | 54µs            | 116µs           | 27126.67    
📋 pod-to-pod      | same-node  | UDP_RR          | 1s              | 14µs            | 30.04µs         | 1.442ms         | 17µs            | 48µs            | 90µs            | 33049.72    
📋 pod-to-pod      | same-node  | TCP_CRR         | 1s              | 286µs           | 490.63µs        | 7.024ms         | 458µs           | 560µs           | 1.125ms         | 2034.74     
📋 pod-to-host     | same-node  | TCP_RR          | 1s              | 12µs            | 31.94µs         | 1.674ms         | 38µs            | 44µs            | 81µs            | 31080.33    
📋 pod-to-host     | same-node  | UDP_RR          | 1s              | 12µs            | 31.07µs         | 1.704ms         | 37µs            | 44µs            | 88µs            | 31974.69    
📋 pod-to-host     | same-node  | TCP_CRR         | 1s              | 190µs           | 333.39µs        | 6.135ms         | 313µs           | 396µs           | 710µs           | 2991.20     
📋 host-to-pod     | same-node  | TCP_RR          | 1s              | 12µs            | 32.51µs         | 7.725ms         | 37µs            | 45µs            | 99µs            | 30553.66    
📋 host-to-pod     | same-node  | UDP_RR          | 1s              | 12µs            | 27.8µs          | 1.226ms         | 22µs            | 42µs            | 83µs            | 35702.75    
📋 host-to-pod     | same-node  | TCP_CRR         | 1s              | 187µs           | 361.23µs        | 9.06ms          | 319µs           | 415µs           | 1.02ms          | 2761.72     
📋 host-to-host    | same-node  | TCP_RR          | 1s              | 12µs            | 30.9µs          | 4.123ms         | 35µs            | 42µs            | 83µs            | 32146.29    
📋 host-to-host    | same-node  | UDP_RR          | 1s              | 11µs            | 28.42µs         | 2.547ms         | 34µs            | 40µs            | 101µs           | 34959.06    
📋 host-to-host    | same-node  | TCP_CRR         | 1s              | 62µs            | 146.65µs        | 8.981ms         | 127µs           | 172µs           | 506µs           | 6794.35     
📋 pod-to-pod      | other-node | TCP_RR          | 1s              | 331µs           | 730.65µs        | 13ms            | 650µs           | 1.037ms         | 2.125ms         | 1363.17     
📋 pod-to-pod      | other-node | UDP_RR          | 1s              | 310µs           | 636.93µs        | 4.923ms         | 581µs           | 891µs           | 1.45ms          | 1564.81     
📋 pod-to-pod      | other-node | TCP_CRR         | 1s              | 936µs           | 2.12513ms       | 9.839ms         | 1.585ms         | 5.566ms         | 6.85ms          | 469.93      
📋 pod-to-host     | other-node | TCP_RR          | 1s              | 243µs           | 529.46µs        | 7.377ms         | 499µs           | 727µs           | 1.058ms         | 1881.60     
📋 pod-to-host     | other-node | UDP_RR          | 1s              | 205µs           | 506.11µs        | 16.028ms        | 414µs           | 684µs           | 2.066ms         | 1969.76     
📋 pod-to-host     | other-node | TCP_CRR         | 1s              | 607µs           | 1.17934ms       | 15.836ms        | 1.085ms         | 1.529ms         | 2.5ms           | 845.96      
📋 host-to-pod     | other-node | TCP_RR          | 1s              | 254µs           | 537.74µs        | 18.134ms        | 481µs           | 736µs           | 1.287ms         | 1851.63     
📋 host-to-pod     | other-node | UDP_RR          | 1s              | 222µs           | 489.01µs        | 4.161ms         | 446µs           | 696µs           | 1.181ms         | 2036.72     
📋 host-to-pod     | other-node | TCP_CRR         | 1s              | 590µs           | 1.11054ms       | 10.26ms         | 1.03ms          | 1.436ms         | 2.3ms           | 898.38      
📋 host-to-host    | other-node | TCP_RR          | 1s              | 223µs           | 503.58µs        | 4.419ms         | 452µs           | 685µs           | 1.45ms          | 1977.50     
📋 host-to-host    | other-node | UDP_RR          | 1s              | 203µs           | 430.1µs         | 4.677ms         | 388µs           | 600µs           | 1.05ms          | 2315.76     
📋 host-to-host    | other-node | TCP_CRR         | 1s              | 529µs           | 949.29µs        | 8.49ms          | 894µs           | 1.252ms         | 1.9ms           | 1049.70     
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------
📋 Scenario        | Node       | Test            | Duration        | Throughput Mb/s 
-------------------------------------------------------------------------------------
📋 pod-to-pod      | same-node  | TCP_STREAM      | 1s              | 605.80       
📋 pod-to-pod      | same-node  | UDP_STREAM      | 1s              | 443.89       
📋 pod-to-host     | same-node  | TCP_STREAM      | 1s              | 1102.85      
📋 pod-to-host     | same-node  | UDP_STREAM      | 1s              | 612.62       
📋 host-to-pod     | same-node  | TCP_STREAM      | 1s              | 1364.10      
📋 host-to-pod     | same-node  | UDP_STREAM      | 1s              | 645.54       
📋 host-to-host    | same-node  | TCP_STREAM      | 1s              | 6928.44      
📋 host-to-host    | same-node  | UDP_STREAM      | 1s              | 493.50       
📋 pod-to-pod      | other-node | TCP_STREAM      | 1s              | 350.29       
📋 pod-to-pod      | other-node | UDP_STREAM      | 1s              | 141.66       
📋 pod-to-host     | other-node | TCP_STREAM      | 1s              | 1401.42      
📋 pod-to-host     | other-node | UDP_STREAM      | 1s              | 169.30       
📋 host-to-pod     | other-node | TCP_STREAM      | 1s              | 3802.40      
📋 host-to-pod     | other-node | UDP_STREAM      | 1s              | 166.53       
📋 host-to-host    | other-node | TCP_STREAM      | 1s              | 3894.40      
📋 host-to-host    | other-node | UDP_STREAM      | 1s              | 202.85       
-------------------------------------------------------------------------------------
```

Fixes https://github.com/cilium/cilium-cli/issues/1111
Fixes https://github.com/cilium/cilium-cli/issues/2114