@darox 2023-12-11
> /cc @darox @learnitall as you worked on the perf test before, I would appreciate your reviews.

Great, I planned to work on refactoring, but you were faster. Planning to have a look at it this week (obviously customer tickets have priority).