@marseel 2023-12-11
I think we can diregard `EKS (tunnel) / EKS (tunnel) Installation and Connectivity Test (classic) (pull_request_target)` failure as it pulls `in-cluster-test-scripts` from main, where it tries to run `cilium connectivity test --perf --perf-duration 1s`