@darox 2023-12-13
> > /cc @darox @learnitall as you worked on the perf test before, I would appreciate your reviews.
> 
> Great, I planned to work on refactoring, but you were faster. Planning to have a look at it this week (obviously customer tickets have priority).

I see you removed me from the reviewers. Makes sense, the others have deeper Go knowledge than me.