@marseel 2023-12-13
Since I removed `pull_request: {}` from workflows I expect that the test will be failing right now -but this is fine, they were passing before with it.