@pchaigno 2023-12-15
I believe this pull request broke main. At least, the following is currently failing for me on main locally and in my rebased PR:
```
$ make && ./cilium connectivity perf --duration 1s
CGO_ENABLED=0 go build  \
	-ldflags "-w -s \
	-X 'github.com/cilium/cilium-cli/cli.Version=v0.15.18'" \
	-o cilium \
	./cmd/cilium
unknown flag: --duration
```