@pchaigno 2023-12-15
Hm, my local clone wasn't on latest main and after pulling it doesn't fail anymore. And apparently GitHub sometimes lies about PRs being up-to-date because I had this:
![Screenshot from 2023-12-15 18-49-28](https://github.com/cilium/cilium-cli/assets/1764210/13520cd3-7a5f-4c22-b732-9c35f416830a)
Yet I just updated the branch after rebasing...

So false alert (and weird GH bug). Sorry for the noise!